move functionality into cleantemp pkg
